Moving App component into index file
