+++ src/style/index.js
@@ -1,0 +1,2 @@
+import GlobalStyle from "./globalStyle"
+export { GlobalStyle }
